+++ lib/segment/analytics/backoff_policy.rb
@@ -1,9 +1,9 @@
 require 'segment/analytics/defaults'
 
-module Segment
+module SegmentAPI
   class Analytics
     class BackoffPolicy
-      include Segment::Analytics::Defaults::BackoffPolicy
+      include SegmentAPI::Analytics::Defaults::BackoffPolicy
 
       # @param [Hash] opts
       # @option opts [Numeric] :min_timeout_ms The minimum backoff timeout