+++ lib/segment/analytics/client.rb
@@ -6,11 +6,11 @@ require 'segment/analytics/logging'
 require 'segment/analytics/utils'
 require 'segment/analytics/worker'
 
-module Segment
+module SegmentAPI
   class Analytics
     class Client
-      include Segment::Analytics::Utils
-      include Segment::Analytics::Logging
+      include SegmentAPI::Analytics::Utils
+      include SegmentAPI::Analytics::Logging
 
       # @param [Hash] opts
       # @option opts [String] :write_key Your project's write_key