+++ spec/segment/analytics/worker_spec.rb
@@ -1,12 +1,12 @@
 require 'spec_helper'
 
-module Segment
+module SegmentAPI
   class Analytics
     describe Worker do
       describe '#init' do
         it 'accepts string keys' do
           queue = Queue.new
-          worker = Segment::Analytics::Worker.new(queue,
+          worker = SegmentAPI::Analytics::Worker.new(queue,
                                                   'secret',
                                                   'batch_size' => 100)
           batch = worker.instance_variable_get(:@batch)
@@ -16,11 +16,11 @@ module Segment
 
       describe '#run' do
         before :all do
-          Segment::Analytics::Defaults::Request::BACKOFF = 0.1
+          SegmentAPI::Analytics::Defaults::Request::BACKOFF = 0.1
         end
 
         after :all do
-          Segment::Analytics::Defaults::Request::BACKOFF = 30.0
+          SegmentAPI::Analytics::Defaults::Request::BACKOFF = 30.0
         end
 
         it 'does not error if the endpoint is unreachable' do
@@ -29,7 +29,7 @@ module Segment
 
             queue = Queue.new
             queue << {}
-            worker = Segment::Analytics::Worker.new(queue, 'secret')
+            worker = SegmentAPI::Analytics::Worker.new(queue, 'secret')
             worker.run
 
             expect(queue).to be_empty
@@ -39,10 +39,10 @@ module Segment
         end
 
         it 'executes the error handler if the request is invalid' do
-          Segment::Analytics::Request
+          SegmentAPI::Analytics::Request
             .any_instance
             .stub(:post)
-            .and_return(Segment::Analytics::Response.new(400, 'Some error'))
+            .and_return(SegmentAPI::Analytics::Response.new(400, 'Some error'))
 
           status = error = nil
           on_error = proc do |yielded_status, yielded_error|
@@ -60,7 +60,7 @@ module Segment
           sleep 0.1 # First give thread time to spin-up.
           sleep 0.01 while worker.is_requesting?
 
-          Segment::Analytics::Request.any_instance.unstub(:post)
+          SegmentAPI::Analytics::Request.any_instance.unstub(:post)
 
           expect(queue).to be_empty
           expect(status).to eq(400)
@@ -111,7 +111,7 @@ module Segment
       describe '#is_requesting?' do
         it 'does not return true if there isn\'t a current batch' do
           queue = Queue.new
-          worker = Segment::Analytics::Worker.new(queue, 'testsecret')
+          worker = SegmentAPI::Analytics::Worker.new(queue, 'testsecret')
 
           expect(worker.is_requesting?).to eq(false)
         end
@@ -119,7 +119,7 @@ module Segment
         it 'returns true if there is a current batch' do
           queue = Queue.new
           queue << Requested::TRACK
-          worker = Segment::Analytics::Worker.new(queue, 'testsecret')
+          worker = SegmentAPI::Analytics::Worker.new(queue, 'testsecret')
 
           worker_thread = Thread.new { worker.run }
           eventually { expect(worker.is_requesting?).to eq(true) }